More appropriate data types (#28)
* use more appropiate data types for "timestamps" across modules

* use unsigned types on depth update_id's

* remove unused deps, update deps & bump `Iced` rev

* chore: remove redundant casting & improve error handling on `bybit` conn.

* fix a crash when overflow on calc. on chart init phase
-- && some tweaks on styling

* chore: styling tweaks

* deserialize orders directly to f32's instead of "String > f32" route

* chore: remove redundant deserializer impl

* deserialize some more structs to desired types directly

* chore: make local depth cache a little more idiomatic

+++ src/charts/timeandsales.rs
@@ -45,7 +45,10 @@ impl TimeAndSales {
 
         for trade in trades_buffer {
             if let Some(trade_time) =
-                DateTime::from_timestamp(trade.time / 1000, (trade.time % 1000) as u32 * 1_000_000)
+                DateTime::from_timestamp(
+                    trade.time as i64 / 1000, 
+                    (trade.time % 1000) as u32 * 1_000_000
+                )
             {
                 let converted_trade = ConvertedTrade {
                     time_str: trade_time.format("%M:%S.%3f").to_string(),
